Revert "Merge https://github.com/hexparrot/mineos-node into waterfall-paper-api"
This reverts commit 9f7ba88159c51d8ff40296ff9f2d1defa4e48b9d, reversing
changes made to f599afaffc284719d26e1608a2fc3dc8aa996e14.
