--- Makefile
@@ -1,20 +1,0 @@
-REGISTRY_NAME := 
-REPOSITORY_NAME := hexparrot/
-IMAGE_NAME := mineos
-TAG := :latest
-
-.PHONY: getcommitid
-all: build
-
-getcommitid: 
-	$(eval COMMITID = $(shell git log -1 --pretty=format:"%H"))
-
-build: getcommitid
-	@docker build -t $(REGISTRY_NAME)$(REPOSITORY_NAME)$(IMAGE_NAME):$(COMMITID) -f Dockerfile .
-	@docker build -t $(REGISTRY_NAME)$(REPOSITORY_NAME)$(IMAGE_NAME)$(TAG) -f Dockerfile .
-
-publish: build
-	docker login
-	docker push $(REGISTRY_NAME)$(REPOSITORY_NAME)$(IMAGE_NAME)$(TAG)
-	docker push $(REGISTRY_NAME)$(REPOSITORY_NAME)$(IMAGE_NAME):$(COMMITID)
-	docker logout
