+++ html/locales/locale-ru_RU.json
@@ -42,8 +42,8 @@
   "PORT": "Порт",
   "STATUS": "Статус",
   "MEMORY": "Память",
-  "UP": "Запущен",
-  "DOWN": "Остановлен",
+  "UP": "Вверх",
+  "DOWN": "Вниз",
 
   "SERVER_NAME": "Имя сервера",
   "CREATE_HELP": "Только латинские буквы, цифры, подчеркивания.",
@@ -77,8 +77,6 @@
   "COPY_TO_SERVER": "Скопировать в сервер",
 
   "GLANCE": "Обзор сервера", 
-  "INVALID_SERVER_NAME": "Имя сервера может содержать только латинские буквы, цифры и подчеркивания. Этим сервером нельзя управлять в MineOS без переименования.",
-  "LOG_TAILING_RATE_LIMITED": "Сервер слишком быстро генерирует протокол (лог). Обновление протокола остановлено - обновите страницу чтобы возобновить.",
   "VERSION": "Версия",
   "REPORTED_VERSION": "Обнаруженная версия",
   "MEMORY_FOOTPRINT": "Занятая память",
@@ -123,7 +121,6 @@
   "ARCHIVE_CREATED": "Архив создан",
   "ARCHIVE_FAILED": "Попытка создания архива неудачна",
   "CREATE_NEW_ARCHIVE": "Создать новый архив",
-  "COMMIT_THEN_CREATE_ARCHIVE": "Сохранить сервер а затем создать архив",
 
   "OWNERSHIP_AND_DISK_USAGE": "Владелец и место на диске",
   "SERVER_OWNER": "Владелец сервера",
@@ -181,9 +178,9 @@
 
   "SUCCEEDED": "успешно",
   "FAILED": "неудачно",
-  "!up": "Сервер не должен работать для выполнении этой команды.",
-  "up": "Сервер должен работать для выполнения этой команды.",
-  "!exists": "Сервер не должен существовать для выполнении этой команды.",
+  "!up": "Сервер мог быть выключен при выполнении этой команды.",
+  "up": "Сервер должен быть запущен для этой команды.",
+  "!exists": "Сервер мог не существовать при выполнении этой команды.",
   "exists": "Не найден сервер для выполнения команды",
   "eula": "Для запуска этого сервера необходимо принять лицензионное соглашение с конечным пользователем (EULA)."
 }
