+++ package.json
@@ -22,13 +22,12 @@
   },
   "dependencies": {
     "adm-zip": "^0.4.11",
-    "angular": "1.5.x",
+    "angular": "^1.4.14",
     "angular-moment": "1.0.x",
     "angular-sanitize": "1.5.x",
     "angular-translate": "^2.18.2",
     "apache-crypt": "1.1.x",
     "async": "1.5.x",
-    "bindings": "1.5.0",
     "body-parser": "^1.19.0",
     "chownr": "1.0.x",
     "compression": "^1.7.2",
@@ -42,7 +41,7 @@
     "express-session": "^1.16.2",
     "fireworm": "0.7.x",
     "fs-extra": "5.0.x",
-    "ini": "^1.3.6",
+    "ini": "^1.3.5",
     "introspect": "0.0.x",
     "lazylines": "1.0.0",
     "line-by-line": "^0.1.6",
@@ -52,7 +51,6 @@
     "moment-duration-format": "1.3.x",
     "moment-timezone": "^0.5.17",
     "node-getopt": "^0.2.4",
-    "node-addon-api": "3.1.0",
     "node-uuid": "1.4.x",
     "object-hash": "1.1.x",
     "passport": "0.3.x",
@@ -65,7 +63,7 @@
     "request-progress": "0.3.x",
     "rsync": "0.6.x",
     "sha512crypt-node": "0.1.x",
-    "socket.io": "1.7.x",
+    "socket.io": "1.5.x",
     "strftime": "0.9.x",
     "tail": "1.2.1",
     "tar": "^2.2.2",
@@ -73,7 +71,7 @@
     "tmp": "0.0.33",
     "unzip": "0.1.x",
     "urllib": "^2.34.1",
-    "userid": "1.0.0-beta.9",
+    "userid": "^0.3.1",
     "uuid": "3.0.x",
     "which": "1.2.x",
     "whoami": "0.0.x",
