--- profiles.d/spongevanilla.js
@@ -1,58 +1,0 @@
-var async = require('async');
-var path = require('path');
-var fs = require('fs-extra');
-var profile = require('./template');
-
-exports.profile = {
-  name: 'SpongeVanilla',
-  request_args: {
-    url: 'https://repo.spongepowered.org/maven/org/spongepowered/spongevanilla/maven-metadata.xml',
-    json: false,
-    gzip: true
-  },
-  handler: function (profile_dir, body, callback) {
-    var p = [];
-
-    try {
-      var xml_parser = require('xml2js');
-
-      xml_parser.parseString(body, function (inner_err, result) {
-        try {
-          var packs = result['metadata']['versioning'][0]['versions'][0]['version'];
-
-          for (var index in packs) {
-            var item = new profile();
-            var matches = packs[index].match(/([\d\.]+)-([\d\.]+)?-?(\D+)-(\d+)/);
-
-            item['version'] = packs[index];
-            item['group'] = 'spongevanilla';
-
-            switch (matches[3]) {
-              case 'DEV':
-                item['type'] = 'snapshot';
-                break;
-              case 'BETA':
-                item['type'] = 'release';
-                break;
-              default:
-                item['type'] = 'old_versions';
-                break;
-            }
-
-            item['id'] = 'SpongeVanilla-{0}{1}{2}'.format(matches[1], matches[3][0].toLowerCase(), matches[4]);
-            item['webui_desc'] = 'Version {0}, build {1} (mc: {2})'.format(matches[2], matches[4], matches[1]);
-            item['weight'] = 5;
-            item['filename'] = 'spongevanilla-{0}.jar'.format(item.version);
-            item['downloaded'] = fs.existsSync(path.join(profile_dir, item.id, item.filename));
-            item['url'] = 'https://repo.spongepowered.org/maven/org/spongepowered/spongevanilla/{0}/spongevanilla-{0}.jar'.format(item.version);
-            p.push(item);
-          }
-          callback(inner_err, p);
-        } catch (e) { }
-      })
-
-    } catch (e) { }
-
-    callback(null, p);
-  } //end handler
-}
